feat: 과정 상세조회 모달 구현(#114)
- 리프레시 토큰 자동 갱신 적용
- 로딩 및 오류 상태 처리 개선
